requirements.txt
contourpy==1.3.0
cycler==0.12.1
fonttools==4.60.2
greenlet==3.2.4
importlib_resources==6.5.2
joblib==1.5.3
kiwisolver==1.4.7
llvmlite==0.43.0
matplotlib==3.9.4
numba==0.60.0
numpy==2.0.2
nvidia-nccl-cu12==2.28.9
packaging==25.0
pandas==2.3.3
pillow==11.3.0
PyMySQL==1.1.2
pyparsing==3.2.5
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pytz==2025.2
scikit-learn==1.6.1
scipy==1.13.1
seaborn==0.13.2
six==1.17.0
SQLAlchemy==2.0.45
threadpoolctl==3.6.0
tqdm==4.67.1
typing_extensions==4.15.0
tzdata==2025.3
xgboost==2.1.4
zipp==3.23.0
